+++ e2e/tests/workspaces/API/onFrameClosed.spec.js
@@ -78,6 +78,7 @@ describe('onFrameClosed() ', () => {
                 try {
                     expect(closed).to.be.an("object");
                     expect(closed.frameId).to.be.a("string");
+                    expect(closed.frameBounds).to.be.an("object");
                     ready();
                 } catch (error) {
                     done(error);
@@ -110,6 +111,32 @@ describe('onFrameClosed() ', () => {
             })
             .then(ready)
             .catch(done);
+    });
+
+    it('should provide the correct frameBounds when notifying of a closed frame', async () => {
+        const wrapper = gtf.wrapPromise();
+        const bounds = {
+            left: 42,
+            top: 84,
+            width: 600,
+            height: 500
+        };
+
+        const workspace = await glue.workspaces.createWorkspace(Object.assign({}, basicConfig, { frame: { newFrame: { bounds } } }));
+
+        const unSub = await glue.workspaces.onFrameClosed((closed) => {
+            try {
+                expect(closed.frameBounds).to.eql(bounds);
+                wrapper.resolve();
+            } catch (error) {
+                wrapper.reject(error);
+            }
+        })
+        unSubFuncs.push(unSub);
+
+        await workspace.frame.close();
+
+        return wrapper.promise;
     });
 
     it('should notify exactly once of a closed frame', (done) => {