+++ e2e/tests/workspaces/API/onWorkspaceClosed.spec.js
@@ -80,6 +80,7 @@ describe('onWorkspaceClosed ', () => {
                 try {
                     expect(closed.frameId).to.be.a('string');
                     expect(closed.workspaceId).to.be.a('string');
+                    expect(closed.frameBounds).to.be.an('object');
                     ready();
                 } catch (error) {
                     done(error);
@@ -114,6 +115,54 @@ describe('onWorkspaceClosed ', () => {
             })
             .then(ready)
             .catch(done);
+    });
+
+    it('should provide the correct frame bounds when notifying of a closed workspace', async () => {
+        const wrapper = gtf.wrapPromise();
+        const bounds = {
+            left: 42,
+            top: 84,
+            width: 400,
+            height: 500
+        };
+        const workspace = await glue.workspaces.createWorkspace(Object.assign({}, basicConfig, { frame: { newFrame: { bounds } } }));
+
+        const unSub = await glue.workspaces.onWorkspaceClosed((data) => {
+            expect(data.frameBounds).to.eql(bounds);
+            wrapper.resolve();
+        });
+
+        unSubFuncs.push(unSub);
+
+        await workspace.close();
+
+        return wrapper.promise;
+    });
+
+    it('should provide the correct frame bounds when closing the frame', async () => {
+        const wrapper = gtf.wrapPromise();
+        const bounds = {
+            left: 42,
+            top: 84,
+            width: 400,
+            height: 500
+        };
+        const workspace = await glue.workspaces.createWorkspace(Object.assign({}, basicConfig, { frame: { newFrame: { bounds } } }));
+
+        const unSub = await glue.workspaces.onWorkspaceClosed((data) => {
+            try {
+                expect(data.frameBounds).to.eql(bounds);
+                wrapper.resolve();
+            } catch (error) {
+                wrapper.reject(error);
+            }
+        });
+
+        unSubFuncs.push(unSub);
+
+        await workspace.frame.close();
+
+        return wrapper.promise;
     });
 
     it('should notify exactly once of a closed workspace by closing the frame', (done) => {