+++ packages/workspaces-api/changelog.md
@@ -1,3 +1,5 @@
+1.16.1
+feat: added frameBounds to the data provided to workspace and frame closed events
 1.16.0
 feat: added support for allowWorkspaceTabExtract
 1.15.5