+++ packages/workspaces-ui-core/changelog.md
@@ -1,3 +1,5 @@
+1.9.2
+feat: started providing frameBounds for frame and workspace closed events
 1.9.1
 chore: bump due to dependencies update
 1.9.0